feat: move code to main
