+++ .github/actions/prepare/action.yml
@@ -5,10 +5,13 @@ name: Prepare
 runs:
   steps:
     - uses: pnpm/action-setup@v2
+
     - uses: actions/setup-node@v4
       with:
         cache: pnpm
         node-version: "20"
+
     - run: pnpm install --frozen-lockfile
       shell: bash
+
   using: composite
