+++ src/index.ts
@@ -1,24 +1,9 @@
-export * from "./validationResult.js";
-import core from "@actions/core";
-import { processValidationResult, validateUrl } from "./validate.js";
+export * from "./validationResult.js"; // not sure if this is useful
 
-try {
-	const url = core.getInput("url");
+/**
+ * The entrypoint for the action.
+ */
+import { run } from "./main.js";
 
-	const validationResult = await validateUrl(url);
-	const processedValidationResult = processValidationResult(validationResult);
-
-	core.setOutput("validationResult", validationResult);
-	core.setOutput("processedValidationResult", processedValidationResult);
-
-	if (processedValidationResult.success) {
-		console.log(processedValidationResult.resultText);
-	} else {
-		core.setFailed(processedValidationResult.resultText);
-	}
-
-	console.log(processedValidationResult.resultText);
-} catch (error) {
-	// eslint-disable-next-line @typescript-eslint/restrict-template-expressions
-	core.setFailed(error instanceof Error ? error.message : `error: ${error}`);
-}
+// eslint-disable-next-line @typescript-eslint/no-floating-promises
+run();
